Refactor(LottoNumberUtil): 클래스명 변경 및 유틸 클래스로 변경
