+++ src/main/java/lotto/util/LottoNumberUtil.java
@@ -1,10 +1,10 @@
-package lotto.model.validator;
+package lotto.util;
 
 import static lotto.model.constant.ErrorMessage.INVALID_LOTTO_NUMBER_RANGE;
 import static lotto.model.constant.LottoConstant.MAX_NUMBER;
 import static lotto.model.constant.LottoConstant.MIN_NUMBER;
 
-public class LottoNumberValidator {
+public class LottoNumberUtil {
 
     public static void validateNumber(int number) {
         if (number < MIN_NUMBER || MAX_NUMBER < number) {
